pyre::typelists: renamed the header with the {types_t} list
