+++ lib/pyre/typelists/append.h
@@ -12,7 +12,7 @@
 #include "forward.h"
 
 // support
-#include "typelist.h"
+#include "types.h"
 
 
 // append the types {T} to an existing {typesT} list
