+++ lib/pyre/typelists/cartesian.h
@@ -12,7 +12,7 @@
 #include "forward.h"
 
 // support
-#include "typelist.h"
+#include "types.h"
 #include "merge.h"
 #include "concat.h"
