+++ lib/pyre/typelists/concat.h
@@ -12,7 +12,7 @@
 #include "forward.h"
 
 // support
-#include "typelist.h"
+#include "types.h"
 
 
 // base case: no arguments -> empty list
