+++ lib/pyre/typelists/merge.h
@@ -12,7 +12,7 @@
 #include "forward.h"
 
 // support
-#include "typelist.h"
+#include "types.h"
 
 
 // type U type
