+++ lib/pyre/typelists/prepend.h
@@ -9,7 +9,7 @@
 #include "forward.h"
 
 // support
-#include "typelist.h"
+#include "types.h"
 
 
 // prepend a type {T} to an existing {typesT} list
