+++ lib/pyre/typelists/templates.h
@@ -1,0 +1,19 @@
+// -*- c++ -*-
+//
+// michael a.g. aïvázis <michael.aivazis@para-sim.com>
+// (c) 1998-2025 all rights reserved
+
+// code guard
+#pragma once
+
+
+// forward declaration
+#include "forward.h"
+
+
+// a type container
+template <template <typename...> class...>
+struct pyre::typelists::templates_t {};
+
+
+// end of file
